+++ RAMS/Mobile/RAMMS.MobileApps/PageModel/FormDAddMaterialPageModel.cs
@@ -50,7 +50,7 @@ namespace RAMMS.MobileApps.PageModel
 
         public string SelectedUnit { get; set; }
 
-        public int? strQty { get; set; }
+        public decimal? strQty { get; set; }
 
         public string strRemarksValue { get; set; }
